- able to choose from the options of intervals, next task is to figure out how to set date and time

+++ LawsForImpact/Models/SavedInformation.cs
@@ -10,6 +10,6 @@ namespace LawsForImpact.Models
         public int QueueIndex { get; set; }
         public int Index { get; set; }
         public bool RandomToggle { get; set; }
-        public int RepeatInterval { get; set; }
+        public long RepeatInterval { get; set; }
     }
 }
